pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "livekit-agents[assemblyai,cartesia,deepgram,google,inworld,mcp,openai,silero,turn-detector]~=1.4",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.2.2",
]
